// src/lit.smabu.react/src/types/domain/currency.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LIT.Smabu.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface Currency
 */
export interface Currency {

    /**
     * @type {string}
     * @memberof Currency
     */
    isoCode?: string | null;

    /**
     * @type {string}
     * @memberof Currency
     */
    name?: string | null;

    /**
     * @type {string}
     * @memberof Currency
     */
    sign?: string | null;
}