// src/lit.smabu.react/src/types/domain/invoice-id.ts
/* tslint:disable */
/* eslint-disable */
/**
 * LIT.Smabu.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * 
 *
 * @export
 * @interface InvoiceId
 */
export interface InvoiceId {

    /**
     * @type {string}
     * @memberof InvoiceId
     */
    value?: string;
}